core/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.10
django-extensions==3.1.5
django-filter==21.1
django-taggit==2.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.12.4
importlib-metadata==4.8.2
Markdown==3.3.6
Pillow==8.4.0
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
zipp==3.6.0
